Card Class and Card List
Started a very basic card object. Not entirely sure of the best way to
define a card - we'll have to brainstorm some ideas. I also added a
spreadsheet that contains the name, type, value, and count of each card.
Counts still need to be added for most cards (non-property cards). This
spreadsheet will be read in and used to create the deck.


git-svn-id: https://github.com/robinschiro/MonopolyDeal/trunk@7 8c082f89-c81d-736b-f694-ce57b7db7bff